ci: Add support for releases
Enhanced build for cross-compiling with a given OS (e.g. compile for X64
and AArch64 on Linux) in a single pass. Added support for platform-
specific source trees.

Built a `release` task to generate platform-specific archives containing
the executable. Added a GitHub workflow to perform a release and
publish artifacts when a `vX.Y.Z` tag is pushed.

Enabled the Gradle config cache.

Refs: #2